refactor: separate large load service into smaller services according to their roles
