+++ internal/core/load/metrics_agent_service.go
@@ -1,0 +1,223 @@
+package load
+
+import (
+	"context"
+	"errors"
+	"fmt"
+	"os"
+	"time"
+
+	"github.com/cloud-barista/cm-ant/internal/infra/outbound/tumblebug"
+	"github.com/cloud-barista/cm-ant/internal/utils"
+)
+
+// InstallMonitoringAgent installs a monitoring agent on specified VMs or all VM on mci.
+func (l *LoadService) InstallMonitoringAgent(param MonitoringAgentInstallationParams) ([]MonitoringAgentInstallationResult, error) {
+	utils.LogInfo("Starting installation of monitoring agent...")
+	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
+	defer cancel()
+	var res []MonitoringAgentInstallationResult
+
+	scriptPath := utils.JoinRootPathWith("/script/install-server-agent.sh")
+	utils.LogInfof("Reading installation script from %s", scriptPath)
+	installScript, err := os.ReadFile(scriptPath)
+	if err != nil {
+		utils.LogErrorf("Failed to read installation script: %v", err)
+		return res, err
+	}
+	username := "cb-user"
+
+	utils.LogInfof("Fetching mci object for NS: %s, msi id: %s", param.NsId, param.MciId)
+	mci, err := l.tumblebugClient.GetMciWithContext(ctx, param.NsId, param.MciId)
+	if err != nil {
+		utils.LogErrorf("Failed to fetch mci : %v", err)
+		return res, err
+	}
+
+	if len(mci.VMs) == 0 {
+		utils.LogErrorf("No VMs found on mci. Provision VM first.")
+		return res, errors.New("there is no vm on mci. provision vm first")
+	}
+
+	var mapSet map[string]struct{}
+	if len(param.VmIds) > 0 {
+		mapSet = utils.SliceToMap(param.VmIds)
+	}
+
+	var errorCollection []error
+
+	for _, vm := range mci.VMs {
+		if mapSet != nil && !utils.Contains(mapSet, vm.ID) {
+			continue
+		}
+		ctx, cancel := context.WithTimeout(ctx, 30*time.Second)
+		defer cancel()
+		m := MonitoringAgentInfo{
+			Username:  username,
+			Status:    "installing",
+			AgentType: "perfmon",
+			NsId:      param.NsId,
+			MciId:     param.MciId,
+			VmId:      vm.ID,
+			VmCount:   len(mci.VMs),
+		}
+		utils.LogInfof("Inserting monitoring agent installation info into database vm id : %s", vm.ID)
+		err = l.loadRepo.InsertMonitoringAgentInfoTx(ctx, &m)
+		if err != nil {
+			utils.LogErrorf("Failed to insert monitoring agent info for vm id %s : %v", vm.ID, err)
+			errorCollection = append(errorCollection, err)
+			continue
+		}
+
+		commandReq := tumblebug.SendCommandReq{
+			Command:  []string{string(installScript)},
+			UserName: username,
+		}
+
+		utils.LogInfof("Sending install command to mci. NS: %s, mci: %s, VMID: %s", param.NsId, param.MciId, vm.ID)
+		_, err = l.tumblebugClient.CommandToVmWithContext(ctx, param.NsId, param.MciId, vm.ID, commandReq)
+
+		if err != nil {
+			if errors.Is(err, context.DeadlineExceeded) {
+				m.Status = "timeout"
+				utils.LogErrorf("Timeout of context. Already 15 seconds has been passed. vm id : %s", vm.ID)
+			} else {
+				m.Status = "failed"
+				utils.LogErrorf("Error occurred during command execution: %v", err)
+			}
+			errorCollection = append(errorCollection, err)
+		} else {
+			m.Status = "completed"
+		}
+
+		l.loadRepo.UpdateAgentInstallInfoStatusTx(ctx, &m)
+
+		r := MonitoringAgentInstallationResult{
+			ID:        m.ID,
+			NsId:      m.NsId,
+			MciId:     m.MciId,
+			VmId:      m.VmId,
+			VmCount:   m.VmCount,
+			Status:    m.Status,
+			Username:  m.Username,
+			AgentType: m.AgentType,
+			CreatedAt: m.CreatedAt,
+			UpdatedAt: m.UpdatedAt,
+		}
+
+		res = append(res, r)
+		utils.LogInfof(
+			"Complete installing monitoring agent on mics: %s, vm: %s",
+			m.MciId,
+			m.VmId,
+		)
+
+		time.Sleep(time.Second)
+	}
+
+	if len(errorCollection) > 0 {
+		return res, fmt.Errorf("multiple errors: %v", errorCollection)
+	}
+
+	return res, nil
+}
+
+func (l *LoadService) GetAllMonitoringAgentInfos(param GetAllMonitoringAgentInfosParam) (GetAllMonitoringAgentInfoResult, error) {
+	var res GetAllMonitoringAgentInfoResult
+	var monitoringAgentInfos []MonitoringAgentInstallationResult
+	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
+	defer cancel()
+
+	utils.LogInfof("GetAllMonitoringAgentInfos called with param: %+v", param)
+	result, totalRows, err := l.loadRepo.GetPagingMonitoringAgentInfosTx(ctx, param)
+
+	if err != nil {
+		utils.LogErrorf("Error fetching monitoring agent infos: %v", err)
+		return res, err
+	}
+
+	utils.LogInfof("Fetched %d monitoring agent infos", len(result))
+
+	for _, monitoringAgentInfo := range result {
+		var r MonitoringAgentInstallationResult
+		r.ID = monitoringAgentInfo.ID
+		r.NsId = monitoringAgentInfo.NsId
+		r.MciId = monitoringAgentInfo.MciId
+		r.VmId = monitoringAgentInfo.VmId
+		r.VmCount = monitoringAgentInfo.VmCount
+		r.Status = monitoringAgentInfo.Status
+		r.Username = monitoringAgentInfo.Username
+		r.AgentType = monitoringAgentInfo.AgentType
+		r.CreatedAt = monitoringAgentInfo.CreatedAt
+		r.UpdatedAt = monitoringAgentInfo.UpdatedAt
+		monitoringAgentInfos = append(monitoringAgentInfos, r)
+	}
+
+	res.MonitoringAgentInfos = monitoringAgentInfos
+	res.TotalRow = totalRows
+
+	return res, nil
+}
+
+// UninstallMonitoringAgent uninstalls a monitoring agent on specified VMs or all VM on Mci.
+// It takes MonitoringAgentInstallationParams as input and returns the number of affected results and any encountered error.
+func (l *LoadService) UninstallMonitoringAgent(param MonitoringAgentInstallationParams) (int64, error) {
+
+	ctx := context.Background()
+	var effectedResults int64
+
+	utils.LogInfo("Starting uninstallation of monitoring agent...")
+	result, err := l.loadRepo.GetAllMonitoringAgentInfosTx(ctx, param)
+
+	if err != nil {
+		utils.LogErrorf("Failed to fetch monitoring agent information: %v", err)
+		return effectedResults, err
+	}
+
+	scriptPath := utils.JoinRootPathWith("/script/remove-server-agent.sh")
+	utils.LogInfof("Reading uninstallation script from %s", scriptPath)
+
+	uninstallPath, err := os.ReadFile(scriptPath)
+	if err != nil {
+		utils.LogErrorf("Failed to read uninstallation script: %v", err)
+		return effectedResults, err
+	}
+
+	username := "cb-user"
+
+	var errorCollection []error
+	for _, monitoringAgentInfo := range result {
+		ctx, cancel := context.WithTimeout(context.Background(), 15*time.Second)
+		defer cancel()
+
+		commandReq := tumblebug.SendCommandReq{
+			Command:  []string{string(uninstallPath)},
+			UserName: username,
+		}
+
+		_, err = l.tumblebugClient.CommandToVmWithContext(ctx, monitoringAgentInfo.NsId, monitoringAgentInfo.MciId, monitoringAgentInfo.VmId, commandReq)
+		if err != nil {
+			errorCollection = append(errorCollection, err)
+			utils.LogErrorf("Failed to uninstall monitoring agent on mci: %s, VM: %s - Error: %v", monitoringAgentInfo.MciId, monitoringAgentInfo.VmId, err)
+			continue
+		}
+
+		err = l.loadRepo.DeleteAgentInstallInfoStatusTx(ctx, &monitoringAgentInfo)
+
+		if err != nil {
+			utils.LogErrorf("Failed to delete agent installation status for mci: %s, VM: %s - Error: %v", monitoringAgentInfo.MciId, monitoringAgentInfo.VmId, err)
+			errorCollection = append(errorCollection, err)
+			continue
+		}
+
+		utils.LogInfof("Successfully uninstalled monitoring agent on mci: %s, VM: %s", monitoringAgentInfo.MciId, monitoringAgentInfo.VmId)
+
+		time.Sleep(time.Second)
+	}
+
+	if len(errorCollection) > 0 {
+		return effectedResults, fmt.Errorf("multiple errors: %v", errorCollection)
+	}
+
+	return effectedResults, nil
+}
